Renamed a few app_manager test files and moved a test
