+++ corehq/apps/app_manager/tests/test_modules.py
@@ -9,8 +9,14 @@ from corehq.apps.app_manager.models import (
     CaseIndex,
     LoadUpdateAction,
     Module,
+    ReportAppConfig,
+    ReportModule,
 )
+from corehq.apps.app_manager.util import purge_report_from_mobile_ucr
+from corehq.apps.userreports.models import ReportConfiguration
+from corehq.util.test_utils import flag_enabled
 from django.test import SimpleTestCase
+from mock import patch
 
 
 class ModuleTests(SimpleTestCase):
@@ -122,3 +128,27 @@ class AdvancedModuleTests(SimpleTestCase):
         self.form.actions.open_cases[-1].repeat_context = "/data/children"
 
         self.assertTrue(self.form.is_registration_form())
+
+
+class ReportModuleTests(SimpleTestCase):
+
+    @flag_enabled('MOBILE_UCR')
+    @patch('dimagi.ext.couchdbkit.Document.get_db')
+    def test_purge_report_from_mobile_ucr(self, get_db):
+        report_config = ReportConfiguration(domain='domain', config_id='foo1')
+        report_config._id = "my_report_config"
+
+        app = Application.new_app('domain', "App")
+        report_module = app.add_module(ReportModule.new_module('Reports', None))
+        report_module.report_configs = [
+            ReportAppConfig(report_id=report_config._id, header={'en': 'CommBugz'}),
+            ReportAppConfig(report_id='other_config_id', header={'en': 'CommBugz'})
+        ]
+        self.assertEqual(len(app.modules[0].report_configs), 2)
+
+        with patch('corehq.apps.app_manager.util.get_apps_in_domain') as get_apps:
+            get_apps.return_value = [app]
+            # this will get called when report_config is deleted
+            purge_report_from_mobile_ucr(report_config)
+
+        self.assertEqual(len(app.modules[0].report_configs), 1)
